# packages/contracts/internal/core-api.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/v1/drivers/{id}": {
      "get": {
        "tags": [
          "driver-controller"
        ],
        "operationId": "get_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DriverResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "driver-controller"
        ],
        "operationId": "update_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DriverResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "driver-controller"
        ],
        "operationId": "delete_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/v1/collection-points/{id}": {
      "put": {
        "tags": [
          "collection-point-controller"
        ],
        "operationId": "update_5",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionPointRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionPointResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/drivers": {
      "get": {
        "tags": [
          "driver-controller"
        ],
        "operationId": "list_7",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseDriverResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "driver-controller"
        ],
        "operationId": "create_4",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DriverResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/drivers/{id}/notify": {
      "post": {
        "tags": [
          "driver-controller"
        ],
        "operationId": "notify",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotifyDriverRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/v1/drivers/{id}/invite": {
      "post": {
        "tags": [
          "driver-controller"
        ],
        "operationId": "invite_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DriverInviteResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/drivers/{id}/assignment": {
      "get": {
        "tags": [
          "driver-controller"
        ],
        "operationId": "activeAssignment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DriverAssignmentResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "driver-controller"
        ],
        "operationId": "assign_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignVehicleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DriverAssignmentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/drivers/{id}/assignment/end": {
      "post": {
        "tags": [
          "driver-controller"
        ],
        "operationId": "endAssignment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/v1/collection-points": {
      "get": {
        "tags": [
          "collection-point-controller"
        ],
        "operationId": "list_9",
        "parameters": [
          {
            "name": "all",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CollectionPointResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "collection-point-controller"
        ],
        "operationId": "create_6",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionPointRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionPointResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/collection-points/{id}/desativar": {
      "post": {
        "tags": [
          "collection-point-controller"
        ],
        "operationId": "desativar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionPointResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/collection-points/{id}/ativar": {
      "post": {
        "tags": [
          "collection-point-controller"
        ],
        "operationId": "ativar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionPointResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DriverRequest": {
        "required": [
          "cnh",
          "name",
          "status"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 150,
            "minLength": 0,
            "type": "string"
          },
          "cnh": {
            "pattern": "\\d{11}",
            "type": "string"
          },
          "phone": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ATIVO",
              "INATIVO"
            ]
          },
          "cnhValidade": {
            "type": "string",
            "format": "date"
          },
          "email": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "DriverResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "cnh": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "cnhValidade": {
            "type": "string",
            "format": "date"
          },
          "email": {
            "type": "string"
          },
          "hasLogin": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CollectionPointRequest": {
        "required": [
          "endereco",
          "lat",
          "lon",
          "nome"
        ],
        "type": "object",
        "properties": {
          "nome": {
            "type": "string"
          },
          "endereco": {
            "type": "string"
          },
          "lat": {
            "type": "number",
            "format": "double"
          },
          "lon": {
            "type": "number",
            "format": "double"
          },
          "janelaInicio": {
            "type": "string",
            "example": "08:00:00"
          },
          "janelaFim": {
            "type": "string",
            "example": "18:00:00"
          }
        }
      },
      "CollectionPointResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nome": {
            "type": "string"
          },
          "endereco": {
            "type": "string"
          },
          "lat": {
            "type": "number",
            "format": "double"
          },
          "lon": {
            "type": "number",
            "format": "double"
          },
          "janelaInicio": {
            "type": "string",
            "example": "08:00:00"
          },
          "janelaFim": {
            "type": "string",
            "example": "18:00:00"
          },
          "posicaoAjustada": {
            "type": "boolean"
          },
          "ativo": {
            "type": "boolean"
          }
        }
      },
...
      "NotifyDriverRequest": {
        "required": [
          "body",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "body": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "DriverInviteResponse": {
        "type": "object",
        "properties": {
          "driverId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "linkUrl": {
            "type": "string"
          }
        }
      },
      "AssignVehicleRequest": {
        "required": [
          "vehicleId"
        ],
        "type": "object",
        "properties": {
          "vehicleId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "DriverAssignmentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "driverId": {
            "type": "string",
            "format": "uuid"
          },
          "vehicleId": {
            "type": "string",
            "format": "uuid"
          },
          "plate": {
            "type": "string"
          },
          "brand": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PageResponseDriverResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DriverResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  }
}